+++ core/src/test/java/com/graphhopper/GraphHopperTest.java
@@ -787,11 +787,11 @@ public class GraphHopperTest {
         Map<String, List<PathDetail>> details = res.getPathDetails();
         assertEquals(1, details.size());
         List<PathDetail> detailList = details.get(Parameters.Details.AVERAGE_SPEED);
-        assertEquals(9, detailList.size());
+        assertEquals(10, detailList.size());
         assertEquals(5.0, detailList.get(0).getValue());
         assertEquals(0, detailList.get(0).getFirst());
         assertEquals(3.0, detailList.get(1).getValue());
-        assertEquals(res.getPoints().size() - 1, detailList.get(8).getLast());
+        assertEquals(res.getPoints().size() - 1, detailList.get(9).getLast());
     }
 
     @Test
@@ -1860,17 +1860,23 @@ public class GraphHopperTest {
                 setOSMFile(MOSCOW).
                 // add profile with turn costs first when no flag encoder is explicitly added
                         setProfiles(
-                        new Profile(profile2).setVehicle(vehicle).setWeighting(weighting).setTurnCosts(true),
+                        new Profile(profile2).setVehicle(vehicle).setWeighting(weighting).setTurnCosts(true).putHint(U_TURN_COSTS, 30),
                         new Profile(profile1).setVehicle(vehicle).setWeighting(weighting).setTurnCosts(false)
                 ).
                 setStoreOnFlush(true);
         hopper.importOrLoad();
 
         GHRequest req = new GHRequest(55.813357, 37.5958585, 55.811042, 37.594689);
+        req.setPathDetails(Arrays.asList("distance", "time")).getHints().putObject("instructions", true);
         req.setProfile("profile_no_turn_costs");
-        assertEquals(400, hopper.route(req).getBest().getDistance(), 1);
+        ResponsePath best = hopper.route(req).getBest();
+        assertEquals(400, best.getDistance(), 1);
+        consistenceCheck(best);
+
         req.setProfile("profile_turn_costs");
-        assertEquals(1044, hopper.route(req).getBest().getDistance(), 1);
+        best = hopper.route(req).getBest();
+        assertEquals(476, best.getDistance(), 1);
+        consistenceCheck(best);
     }
 
     @Test
@@ -2225,20 +2231,22 @@ public class GraphHopperTest {
         List<PathDetail> speeds = path.getPathDetails().get("max_speed");
         assertDetail(speeds.get(0), "null [0, 4]");
         assertDetail(speeds.get(1), "70.0 [4, 6]");
-        assertDetail(speeds.get(2), "100.0 [6, 31]");
-        assertDetail(speeds.get(3), "80.0 [31, 32]");
-        assertDetail(speeds.get(4), "null [32, 37]");
-        assertDetail(speeds.get(5), "50.0 [37, 38]");
-        assertDetail(speeds.get(6), "null [38, 40]");
+        assertDetail(speeds.get(2), "100.0 [6, 16]");
+        assertDetail(speeds.get(3), "100.0 [16, 31]"); // we do not merge path details at via points
+        assertDetail(speeds.get(4), "80.0 [31, 32]");
+        assertDetail(speeds.get(5), "null [32, 37]");
+        assertDetail(speeds.get(6), "50.0 [37, 38]");
+        assertDetail(speeds.get(7), "null [38, 40]");
 
         // check street_names
         List<PathDetail> streetNames = path.getPathDetails().get("street_ref");
         assertDetail(streetNames.get(0), "KU 11 [0, 4]");
-        assertDetail(streetNames.get(1), "B 85 [4, 32]");
-        assertDetail(streetNames.get(2), "null [32, 34]");
-        assertDetail(streetNames.get(3), "KU 18 [34, 37]");
-        assertDetail(streetNames.get(4), "St 2189 [37, 38]");
-        assertDetail(streetNames.get(5), "null [38, 40]");
+        assertDetail(streetNames.get(1), "B 85 [4, 16]");
+        assertDetail(streetNames.get(2), "B 85 [16, 32]");
+        assertDetail(streetNames.get(3), "null [32, 34]");
+        assertDetail(streetNames.get(4), "KU 18 [34, 37]");
+        assertDetail(streetNames.get(5), "St 2189 [37, 38]");
+        assertDetail(streetNames.get(6), "null [38, 40]");
     }
 
     private void assertInstruction(Instruction instruction, String expectedRef, String expectedInterval, int expectedLength, int expectedPoints) {
@@ -2533,6 +2541,55 @@ public class GraphHopperTest {
     }
 
     @Test
+    void timeDetailBug() {
+        final String profile = "profile";
+        GraphHopper hopper = new GraphHopper()
+                .setProfiles(new Profile(profile).setVehicle("car").setWeighting("fastest").setTurnCosts(true).putHint(U_TURN_COSTS, 80))
+                .setGraphHopperLocation(GH_LOCATION)
+                .setMinNetworkSize(200)
+                .setOSMFile(BAYREUTH);
+        hopper.importOrLoad();
+        GHRequest request = new GHRequest(Arrays.asList(
+                new GHPoint(50.020838, 11.494918),
+                new GHPoint(50.024795, 11.498973),
+                new GHPoint(50.023141, 11.496441)));
+        request.setProfile(profile);
+        request.getHints().putObject("instructions", true);
+        request.setPathDetails(Arrays.asList("distance", "time"));
+        GHResponse response = hopper.route(request);
+        assertFalse(response.hasErrors(), response.getErrors().toString());
+
+        consistenceCheck(response.getBest());
+    }
+
+    private void consistenceCheck(ResponsePath path) {
+        double distance = path.getDistance();
+        long time = path.getTime();
+
+        double instructionDistance = 0;
+        long instructionTime = 0;
+        for (Instruction i : path.getInstructions()) {
+            instructionDistance += i.getDistance();
+            instructionTime += i.getTime();
+        }
+
+        assertEquals(time, instructionTime);
+        assertEquals(distance, instructionDistance, 1e-3);
+
+        double pathDetailDistance = 0;
+        for (PathDetail pd : path.getPathDetails().get("distance")) {
+            pathDetailDistance += (Double) pd.getValue();
+        }
+        assertEquals(distance, pathDetailDistance, 1e-3);
+
+        long pathDetailTime = 0;
+        for (PathDetail pd : path.getPathDetails().get("time")) {
+            pathDetailTime += (Long) pd.getValue();
+        }
+        assertEquals(time, pathDetailTime);
+    }
+
+    @Test
     public void testLoadGraph_implicitEncodedValues_issue1862() {
         GraphHopper hopper = new GraphHopper()
                 .setProfiles(
@@ -2595,3 +2652,4 @@ public class GraphHopperTest {
     }
 
 }
+