+++ reader-gtfs/src/main/java/com/graphhopper/gtfs/analysis/PtGraphAsAdjacencyList.java
@@ -5,12 +5,13 @@ import com.graphhopper.routing.ev.*;
 import com.graphhopper.routing.util.AllEdgesIterator;
 import com.graphhopper.routing.util.EdgeFilter;
 import com.graphhopper.routing.weighting.Weighting;
+import com.graphhopper.search.EdgeKVStorage;
 import com.graphhopper.storage.*;
 import com.graphhopper.util.*;
 import com.graphhopper.util.shapes.BBox;
 
 import java.util.Iterator;
-import java.util.Map;
+import java.util.List;
 
 class PtGraphAsAdjacencyList implements Graph {
     private final PtGraph ptGraph;
@@ -303,12 +304,12 @@ class PtGraphAsAdjacencyList implements Graph {
             }
 
             @Override
-            public EdgeIteratorState setKeyValues(Map<String, Object> map) {
+            public EdgeIteratorState setKeyValues(List<EdgeKVStorage.KeyValue> list) {
                 throw new RuntimeException();
             }
 
             @Override
-            public Map<String, Object> getKeyValues() {
+            public List<EdgeKVStorage.KeyValue> getKeyValues() {
                 throw new RuntimeException();
             }