+++ web/src/test/java/com/graphhopper/application/resources/ExtendedJsonResponseTest.java
@@ -25,6 +25,7 @@ import com.graphhopper.matching.Observation;
 import com.graphhopper.matching.State;
 import com.graphhopper.resources.MapMatchingResource;
 import com.graphhopper.routing.querygraph.VirtualEdgeIteratorState;
+import com.graphhopper.search.EdgeKVStorage;
 import com.graphhopper.storage.IntsRef;
 import com.graphhopper.storage.index.Snap;
 import com.graphhopper.util.EdgeIteratorState;
@@ -88,7 +89,7 @@ public class ExtendedJsonResponseTest {
         pointList.add(-3.4445, -38.9990);
         pointList.add(-3.5550, -38.7990);
         return new VirtualEdgeIteratorState(0, 0, 0, 1, 10, new IntsRef(1),
-                Collections.singletonMap("name", "test of iterator"), pointList, false);
+                EdgeKVStorage.KeyValue.createKV("name", "test of iterator"), pointList, false);
     }
 
 }